requirementscnn.txt
streamlit
torch
torchvision
pandas
numpy
tqdm
pillow
scipy
opencv-python
